Added continuation support.
All evaluations now take place in C stack depth independent of LISP
stack depth. The tl_interp object has two stack fields (both regular
tl_object pairs), a value stack (which holds evaluated expressions) and
a "continuation" stack which holds a stack of "continuations"
(representing mostly calls scheduled for the future). tl_push_apply
pushes a continuation; continuations are normally popped by
tl_apply_next in a tight loop in the driver (main here) using
tl_run_until_done.

C code can no longer force an evaluation of a term (at least reliably).
tl_push_eval attempts to evaluate an expression in an environment and
pushes the result to the value stack; if it returns true, the actual
push will be deferred. Assuming proper stack balance, pushing a
continuation using tl_push_apply which consumes one value *before*
tl_push_eval will cause that callable to be called with the produced
value. This is the only robust way to capture an evaluation.

For convenience, tl_eval_and_then does the above tl_push_apply followed
by tl_push_eval. Also, tl_eval_all_args evaluates an entire list and
applies that to the C continuation (as if mapping eval).

C functions and continuations are the same object. All C functions now
take a void * parameter which is stored in their tl_object, and which is
usually provided when creating a new continuation with tl_new_then to
pass data from an invocation to its continuation. Although no structure
is imposed, it is typical to treat the pointer as a tl_object *. For
"regular" C functions created with tl_new_cfunc, the void * will
consistently be NULL on invocation.

C functions return no value (void). Passing values back from a C
function invocation must be done using the value stack. The simplest way
to do so is with tl_cfunc_return(in, val) which pushes val and returns
from the function. Additionally, a function may call tl_eval_and_then or
tl_eval_all_args (and return manually if desired), which pushes no
values but allows the return to the tl_apply_next trampoline to evaluate
expressions and call a continuation.

Continuation objects are added. These capture the state of the
environment, continuation stack and value stack; when called, they
restore all states and push the evaluation of their first argument.
call-with-current-continuation is implemented as creating one of these
objects with the state gathered from the invokation of call/cc, and thus
the value pushed by invoking the continuation is interpreted as the
return value of call/cc.

Some code cleaning is in order before a mainline merge. In particular,
some signalling "len" values in the cont stack are dubious, as are a
proliferation of comments. The project is now at ~1200SLOC, which
demands some simplification.