C conts are GC-safe, occasional running GC, and Makefile fixes
In order to make C continuations GC-safe, the state has to be assumed to
be a valid tl_object *, resulting in another rather large API change.
Types aside, the effect on code was actually quite minimal, since most C
continuations were passed tl_object * through void * anyway.

After inserting some back-of-the-envelope numbers for running tl_gc on
occasion when pushing a continuation, memory usage of TL is now constant
in *any* tail call recursion, thanks to the design of the continuation
stack itself.

run is now a .PHONY target.