Value stack tagging and "by value" cfuncs
This is another large breaking change, sorry.

Some oddity occurs when cfuncs produce values consumed by other cfuncs;
a notorious example is eval-in&, which, if invoked with ''foo on
display, for example, would attempt to run (display foo), and thus
attempt to evaluate foo for a *third* time. This causes an explosion in
the amount of quoting, and--worse--a need to anticipate how many layers
of eval an expression will pass through.

The root cause of the problem was a prior blindness to whether or not a
value on the value stack was put there as a result of an apply parse, or
as the result of a push from, e.g., a cfunction. By differentiating
between the two, the previous example can be fixed without a superfluous
quote, by observing that the sym "foo" which the eval-in& produces is
meant to be passed directly to display, not evaluated again.

This change requires a dichotomy in C functions--those that accept all
arguments strictly by value (like display), and those that accept
by-name arguments (like define). This dichotomy is analogous to the
func/macro dichotomy, respectively, and much of the existing code was
leveraged to achieve this result. As it turns out, a stark majority of C
functions are by value.

Value stack entries are no longer values; they are pairs, with the first
a value, and the next a (TL) bool. The bool is true if the argument is
"syntactic"--id est, a "name" in the call-by-name scheme--and false if
the argument is "direct"--id est, strictly a value. When evaluating a
func or a by-value cfunc, TL will push_eval to convert a syntactic value
to a direct value. The inverse is not feasible, and will result in an
error.

Miscellany: Continuation object calls are considered to be by-value in
this scheme. Also, because they may accept either syntactic or direct
values, "thens" (C continuations) are now distinguished from cfuncs and
cfunc_byvals, and are not subject to the value class checking above. The
semantics of tl_eval_all_args has also changed (it expects args to be a
list of value cells, as on the value stack), but this should be less
important now that by-value cfuncs exist--the call within eval.c should
suffice for most embedding/extension needs.