Added module support proper
Modules come in two flavors: built-in and dynamic. The default is
dynamic, but they have to be built-in if static linkage is requested
(implicit if minilibc is in use).

Dynamic modules can be loaded with the built-in cfunc tl-load-mod, given
a symbol which is interpreted as a path (quote is generally useful
here). Only when a dynamic module is loaded is its TL_INIT_FUNC called.

Static modules are loaded with the program, as soon as the interpreter
starts. Their TL_INIT_FUNCs are called by main; presently, the order
should be assumed to be arbitrary. Sufficiently interested implementors
should be aware that the function pointers are stored sequentially in a
special .rodata.tm_bmcons ELF section, and can thus be reordered if
desired.

In either case, modules may use all functions in tinylisp.h (they are
extern for building a dynamic module, to be linked against the main
interpreter), and the difference between dynamic and static modules
should be mostly transparent (save initialization time). Although this
isn't necessary, modules are generally intended to introduce new cfuncs
into the top or current environment frame (e.g., with tl_env_set_*).
Module line counts don't count against TinyLISP's core line count.